[multi vector] Use 16-bit products in the MinMax Scalar kernel
B holds nibbles, so four u8 x u4 products sum to at most 15300. The Scalar dot product can multiply in u16 and widen once when accumulating into u32. On a Xeon Platinum 8370C this makes the Scalar kernel 2.6-2.7x faster in a baseline x86-64 build and 2.1-2.2x faster with x86-64-v3.

+++ diskann-quantization/src/matrix_kernels/maxsim/minmax8_x_minmax4.rs
@@ -437,8 +437,8 @@ trait ExtraWide<const PACK: usize, const MR: usize>: Copy {
 }
 
 impl ExtraWide<4, 8> for Scalar {
-    type Query = [[u32; 8]; 4];
-    type Splat = [u32; 4];
+    type Query = [[u16; 8]; 4];
+    type Splat = [u16; 4];
     type Accumulator = [u32; 8];
 
     #[cfg(test)]
@@ -449,7 +449,7 @@ impl ExtraWide<4, 8> for Scalar {
     #[inline(always)]
     fn load(self, values: packed::Patch<'_, u8, 8, 4>, _: usize) -> Self::Query {
         let values = values.as_array();
-        core::array::from_fn(|d| core::array::from_fn(|row| u32::from(values[row][d])))
+        core::array::from_fn(|d| core::array::from_fn(|row| u16::from(values[row][d])))
     }
     #[inline(always)]
     fn zero(self) -> Self::Accumulator {
@@ -457,13 +457,14 @@ impl ExtraWide<4, 8> for Scalar {
     }
     #[inline(always)]
     fn splat(self, value: [u8; 4]) -> Self::Splat {
-        value.map(u32::from)
+        value.map(u16::from)
     }
     #[inline(always)]
     fn dot(self, a: Self::Query, b: Self::Splat, acc: Self::Accumulator) -> Self::Accumulator {
+        // B holds nibbles, so four u8 x u4 products sum to at most 15300 and fit in u16.
         core::array::from_fn(|i| {
             let dot = a[0][i] * b[0] + a[1][i] * b[1] + a[2][i] * b[2] + a[3][i] * b[3];
-            acc[i].wrapping_add(dot)
+            acc[i].wrapping_add(u32::from(dot))
         })
     }
